Update on work between plcComm object and Interfaz to work with import such object
